Add an rest-active-old-version css class.
This will be for active versions that are still supported,
but not the current release.

This will be used for our version comparison code.
For example,
When 2.1 is released, it will be green,
2.0 will be yellow,
1.0 will be red.

This will provide an obvious color coded deprecation path.

We should probably add more context around the color so
people know what exactly it means, but this is a start.