[FEATURE] Improve docker build setup
By moving the build step into a separate stage we have a smaller docker
images which speeds up the downloads of the docker image.
Therefor I also switched to the alpine image to get as small as possible.

More changes can be added later, when we want to reduce the size even more.