[FEATURE] Improve docker build setup (#5)
By moving the build step into a separate stage we have a smaller docker
images which speeds up the downloads of the docker image. Therefor I
also switched to the alpine image to get as small as possible.

I added a workflow that will build 3 docker images for different cpu
architectures. Amd64, what most servers and laptops are using nowdays.
Arm/7 for the odd people using a raspberry pi to run their build
pipeline. And arm64 for Mac users. The new MacBooks will take advantage
of this last version. As the os doesn't have to virtualize the amd64. So
docker can be executed natively on the cpu.

More changes can be added later, when we want to reduce the size even
more.