[TASK] relocate config to Documentation (#56)
The default location for guides.xml is Documentation, people might
decide to do it different, so make it configurable.

I update the entrypoint to fix an issue when the input argument was
omitted. The previous implementation did quote all arguments. But that
broke the optional input argument as it would always be `""` when left
empty.

While testing this behavior I found another issue, that was giving some
odd errors when a user does not specify a volume, we to now exit with an
error message.