[DOCS] Document "docker run" issues with macOS, where "--user" is required (#94)
Docker on macOS runs as root. If the entrypoint.sh is then executed, it
sees that the `/project` directory is also mounted as root.

In that case, the script fails because it seems to expect that
`/project` should belong to a specific user. However, the `Dockerfile`
does currently NOT fix the workspace to a user like `typo3`.

Thus, the script only works if either:
* --user is specified or
* docker is automatically run with the current user (seems to be the
case in Linux+Windows)

This PR addresses to document this behaviour.