[TASK] Use php-cs-fixer directly
This change removes the package "typo3/coding-standards" in favour of
"friendsofphp/php-cs-fixer". There is no need to use this proxy package.

Instead, we define the needed rules directly. As we stick zu the PER-CS standard
(the successor of PSR-12) we pin the rules to the PER-CS 1 ruleset. Additionally, we opt in for
some already implemented PER-CS 2 rules to avoid too much style changes already available
in the codebase.

With the pinning to defined rules we ensure that updates of php-cs-fixer does not "break"
the currently used styles locally or in CI. With the ongoing implementation of PER-CS 2 in php-cs-fixer,
this does not "break" anymore the current styles on updates. Changes in the defined rules can therefore be
done explicitly.