[TASK] Adjust classes for various code textroles
Previously, the styles were applies to the code tag with classes. This has the drawback that this
infiltrates all other code tag styles which is undesired. Therefore, a dedicated class "code-inline"
has been introduced to mitigate this.

Additionally, the previous "codeblock.scss" was renamed to "_code.scss" and integrated into the
theme.scss (as the styles has to be loaded anyway for every page as there can be code textroles or
code block directives).