[TASK] Implement "latest" docker container usage, adjust and document workflows/actions (#181)
This PR addresses changes in the process of building docker images, and
also "baking in" a stable version for CDN asset urls.

The general changes are:

* We only use the GIT `main` branch, like before.
* PRs workflow continues with custom GIT branches, like before. 
* PRs will no longer build docker images or the PHAR. These will be
automatically built.
* PRs will still execute the unit and integration test workflows
* Now the docker image will only be built, when we push a version tag to
GIT (i.e. `git tag 1.0.1 && git push --tags`).
* A docker-build test-image will also be built upon changes to the
`Dockerfile`. That image will not be uploaded.
* PHAR generation only on merge to "main" and on releases

For a description of the full workflow, please see the
`Documentation/Developer/Building.rst` with changes included in this PR.

(Please note I had no time to spellcheck/proof-read that or verify that
the RST markup works)