[TASK] Remove doctools.js
This library was mostly responsible for the search in the document.
Usages of the other functions unsuspectible of search are not found.

Additionally, this file was not processed by Grunt, so it always fails
with a 404 when loading a page as it was not available since switching
from the old Sphinx theme cloud URLs.