[TASK] Split up custom JavaScript files (#220)
The code was moved without modification from the previous `theme.js`
file. This makes it easier to get the purpose of a file. They are
combined again to `theme.min.js` upon build.

The initialization of `window.T3Docs = {};` was removed as this is not
used anymore (as far as I can tell).