[CI] Let dependabot create PRs for updating dependencies (#236)
Composer and npm updates can be created automatically using dependabot.
Updates are checked on a daily basis at 16:00. For each pull request
created by dependabot the labels "dependencies", and "composer" (for a
Composer-related update) and "npm" for a (npm-related update) are added.

Development branched are also considered for Composer-related updates
for the time being.

Updates of requirements can also be done manually at any time.