[BUGFIX] Override link handling
Make links in html always relative and in singlehtml
links to documents stay within the document.
Links to resources are always relative

All tests are changed as before now the default
was absolute links, which we overrode in the guides.xml. However absolute links are not usefull for our scenario anyway so I dropped support of them. This does not change the behaviour of our rendered documents

resolves https://github.com/TYPO3-Documentation/render-guides/issues/238