[TASK] composer update phpdocumentor/guides
Additionally phpstan was updated.

The most recent PR of the guides fixed some problems but created new ones.

Resolves https://github.com/TYPO3-Documentation/render-guides/issues/103

I added new tests for the use case toc menu with maxdepth without titlesonly, subpages having subsection.

For some reason up until now the menu was not displayed in the edit-on-gitghub test, now it is correctly displayed.

However in glob toc trees with multiple glob patterns the order is currently not preseved, making that test fail. We are working on a flollow up that will hopefully fix this issue.