[TASK] Add rendertest example
Based on https://typo3-documentation.github.io/sphinx_typo3_theme_rendering_test/Admonitions-and-buttons/Index.html

See Documentation-rendertest/Index.rst on how to build this:

```
make rendertest
composer make rendertest
```

The idea is to be able to preview CSS changes in this directory tree,
distinct from normal Documentation/

After merging this, a first step would be to remove the RST files
with markup that we do no longer want.

The second step then is to verify the remaining RST files for
their proper rendering.