[BUGFIX] Raise docker memory limit to 4G
The Changelog rendering takes up a lot of memory, and thus cannot
be done with the current 1G limit.

A change in guides:1.1.0 introduced caching, and that cache
takes up a lot of space.

Native/local rendering often has no PHP memory limit set, so
this may not apply to CI/development.