[BUGFIX] Use simple PHP container for rendertest and test-docs (#449)
Currently `rendertest` uses the `t3docs:local` docker image, which first
has to be built. But actually it can run with the normal `php` docker
container, just like `make docs` does.

So this removes a dependeny on building a local docker image before and
getting errors.

The same is applied to `make testdocs`, as this should also not require
a local container beforehand.