[TASK] Composer update
Due to guides update the links to confvals are now prefixed with "confval-" as was done in Sphinx. This prevents collisions and keeps links from before the rendering swap the same.

A `:noindex:` option is now available int that case there is no warning about duplicate link targets, but `:confval:` references to the object do not work.

For this we need some template changes that I will do in a follow up so that the link option is not shown when not available.

  - Upgrading league/uri-interfaces (7.4.0 => 7.4.1): Extracting archive
  - Upgrading league/uri (7.4.0 => 7.4.1): Extracting archive
  - Upgrading phpdocumentor/guides (1.2.2 => 1.2.3): Extracting archive
  - Upgrading phpdocumentor/guides-restructured-text (1.2.1 => 1.2.3): Extracting archive
  - Upgrading phpstan/phpstan (1.10.64 => 1.10.65): Extracting archive
  - Upgrading sebastian/environment (6.0.1 => 6.1.0): Extracting archive