[TASK] Add "Shortlink" feature to docs.typo3.org/permalink/
With https://github.com/TYPO3GmbH/site-intercept/pull/188 we
introduced the possibility on docs.typo3.org to forward
to interlink shortcodes to the appropriate targets.

This server-side script is able to go through all
`objects.inv.json` files hosted on the server to parse
the proper targets, and then do a HTTP forward to
the actual documentation files.

This patch implements a new entry block "Shortlink"
(because "permalink" is the actual link to the site,
and still has value to be displayed as a primary
entry) in the popup modal when clicking on a header
reference.

NOTE: This should be regarded as an "experimental"
feature. There may be situations in which the
resolution of permalink targets with confvals and
filenames and others may not work properly yet. We
will need to catch these issues and address
them with follow-up patches to either this
repository or the site-intercept repository
using specific examples.