[TASK] Overwrite run command
Other than what we did before we do now a full overwrite of the run command.
This allows us to remove a number of less optimal constructions in the command.
The change itself is empowered by the changes in phpdocumentor/guides-cli 1.9 which
now provides a number of helper classes to simplify self build run commands.