[FEAT] add watch option
The new `--watch` flag on the run command will help documentation
writers to have a "live" rerendering of their documentation. The
command will run the normal rendering once, then changes to existing
files are detected and will trigger a rerender of the changed file.

To speed up the rendering only the changed file is compiled, this
might cause issues in the menu. But that should not be an issue as
the final result is always based on a full render.