Add :zoom: option support to figure directive
This commit adds support for the :zoom: option on figure directives,
allowing documentation authors to specify different zoom modes for
images.

Supported zoom values:
- lightbox (default) - Dialog overlay
- gallery - PhotoSwipe-style with wheel zoom
- inline - In-place zoom
- lens - Magnifier lens
- css-only - No JS fallback
- none or false - Disable zoom

Example usage:
.. figure:: /Images/screenshot.png
   :zoom: lightbox

   Caption text.

Changes:
- Added custom FigureDirective in T3Docs\Typo3DocsTheme\Directives
- Registered FigureDirective in typo3-docs-theme.php config
- Updated figure.html.twig template to output data-zoom attribute
- FigureNode now receives and stores the zoom option

The implementation extends phpDocumentor's figure directive by adding
the zoom option to the FigureNode, which is then rendered as a
data-zoom attribute in the HTML output.